Allow for managing a single project.
Adds that project to the list of managed projects.

Refactor script into functions.
Create functions for abstracting version control commands for hg and git. This
allows for running 'update' or 'clone' and simply passing in the vcs type,
which will in the future allow for having a file of managed git repos as well
as hg repos.